match6

+++ control-flow/match_case_calculator.py
@@ -1,5 +1,5 @@
-num1= int(input("Enter the first number:"))
-num2= int(input("Enter the second number:"))
+num1= float(input("Enter the first number:"))
+num2= float(input("Enter the second number:"))
 operation = str(input("Choose the operation (+, -, *, /):"))
 
 match operation:
@@ -17,5 +17,4 @@ match operation:
           print("Cannot divide by zero.")
      else:
          result=num1/num2
-         
 print (f"The result is {result}.")
